+++ examples/lpp/location_information.cpp
@@ -9,8 +9,9 @@
 #include "options.hpp"
 #include "utility/types.h"
 
-bool   gConvertConfidence95To39      = false;
+bool   gConvertConfidence95To68      = false;
 double gOverrideHorizontalConfidence = -1;
+bool   gOutputEllipse68              = false;
 
 using namespace location_information;
 
@@ -19,24 +20,19 @@ PLI_Result provide_location_information_callback(UNUSED LocationInformation& loc
                                                  UNUSED void*                userdata) {
 #if 0
     // Example implementation
-    location.tai_time      = TAI_Time::now();
-    location.latitude  = 20;
-    location.longitude = 25;
-    location.altitude = 30;
-    location.bearing = 132;
-    location.horizontal_accuracy = 1;
-    location.horizontal_speed = 2;
-    location.horizontal_speed_accuracy = 3;
-    location.vertical_accuracy = 1;
-    location.vertical_speed = 2;
-    location.vertical_speed_accuracy = 3;
-    location.vertical_velocity_direction = VerticalDirection::UP;
+    location.time     = TAI_Time::now();
+    location.location = LocationShape::ha_ellipsoid_altitude_with_uncertainty(
+        20, 25, 30, HorizontalAccuracy::to_ellipse_39(0.5, 0.5, 0),
+        VerticalAccuracy::from_1sigma(0.5));
 
-    metrics.fixq = FixQuality::RTK_FIX;
-    metrics.sats = 30;
-    metrics.age = 5;
-    metrics.hdop = 10;
-    metrics.vdop = 15;
+    location.velocity = VelocityShape::horizontal_vertical_with_uncertainty(10, 0.5, 90, 1, 0.5,
+                                                                            VerticalDirection::Up);
+
+    metrics.fix_quality          = FixQuality::RTK_FIX;
+    metrics.number_of_satellites = 30;
+    metrics.age_of_corrections   = 5;
+    metrics.hdop                 = 10;
+    metrics.pdop                 = 15;
     return PLI_Result::LI_AND_METRICS;
 #else
     return PLI_Result::NOT_AVAILABLE;
@@ -58,14 +54,26 @@ PLI_Result provide_location_information_callback_ublox(UNUSED LocationInformatio
         return PLI_Result::NOT_AVAILABLE;
     }
 
-    auto semi_major = nav_pvt->h_acc();
-    auto semi_minor = nav_pvt->h_acc();
-    if (gConvertConfidence95To39) {
+    // NOTE(ewasjon): We need to divide the horizontal accuracy by sqrt(2) to get the semi-major and
+    // semi-minor axes. We assume that the semi-major and semi-minor axes are equal.
+    // h_acc = sqrt(semi_major^2 + semi_minor^2)
+    // h_acc = sqrt(2 * semi_major^2) => semi_major = h_acc / sqrt(2)
+    auto semi_major = nav_pvt->h_acc() / 1.4142135623730951;
+    auto semi_minor = nav_pvt->h_acc() / 1.4142135623730951;
+
+    // TODO(ewasjon): Is this really relavant for UBX-NAV-PVT?
+    if (gConvertConfidence95To68) {
+        // 95% confidence to 68% confidence
+        // TODO(ewasjon): should this not be 1.95996 (sqrt(3.84)) from 1-degree chi-squared
+        // distribution?
         semi_major = semi_major / 2.4477;
         semi_minor = semi_minor / 2.4477;
     }
 
-    auto horizontal_accuracy = HorizontalAccuracy::from_ellipse(semi_major, semi_minor, 0);
+    auto horizontal_accuracy = HorizontalAccuracy::to_ellipse_39(semi_major, semi_minor, 0);
+    if (gOutputEllipse68) {
+        horizontal_accuracy = HorizontalAccuracy::to_ellipse_68(semi_major, semi_minor, 0);
+    }
     if (gOverrideHorizontalConfidence >= 0.0) {
         horizontal_accuracy.confidence = gOverrideHorizontalConfidence;
     }
@@ -113,20 +121,31 @@ PLI_Result provide_location_information_callback_nmea(LocationInformation& locat
     auto gga = receiver->gga();
     auto vtg = receiver->vtg();
     auto gst = receiver->gst();
-    if (!gga || !vtg || !gst) {
+    auto epe = receiver->epe();
+    if (!gga || !vtg || !(gst || epe)) {
         return PLI_Result::NOT_AVAILABLE;
     }
 
-    auto semi_major = gst->semi_major();
-    auto semi_minor = gst->semi_minor();
+    auto semi_major  = gst ? gst->semi_major() : epe->semi_major();
+    auto semi_minor  = gst ? gst->semi_minor() : epe->semi_minor();
+    auto orientation = gst ? gst->orientation() : epe->orientation();
+    auto vertical_position_error =
+        gst ? gst->vertical_position_error() : epe->vertical_position_error();
 
-    if (gConvertConfidence95To39) {
+    if (gConvertConfidence95To68) {
+        // 95% confidence to 68% confidence
+        // TODO(ewasjon): should this not be 1.95996 (sqrt(3.84)) from 1-degree chi-squared
+        // distribution?
         semi_major = semi_major / 2.4477;
         semi_minor = semi_minor / 2.4477;
     }
 
     auto horizontal_accuracy =
-        HorizontalAccuracy::from_ellipse(semi_major, semi_minor, gst->orientation());
+        HorizontalAccuracy::to_ellipse_39(semi_major, semi_minor, orientation);
+    if (gOutputEllipse68) {
+        horizontal_accuracy =
+            HorizontalAccuracy::to_ellipse_68(semi_major, semi_minor, orientation);
+    }
     if (gOverrideHorizontalConfidence >= 0.0) {
         horizontal_accuracy.confidence = gOverrideHorizontalConfidence;
     }
@@ -134,7 +153,7 @@ PLI_Result provide_location_information_callback_nmea(LocationInformation& locat
     location.time     = gga->time_of_day();
     location.location = LocationShape::ha_ellipsoid_altitude_with_uncertainty(
         gga->latitude(), gga->longitude(), gga->altitude(), horizontal_accuracy,
-        VerticalAccuracy::from_1sigma(gst->vertical_position_error()));
+        VerticalAccuracy::from_1sigma(vertical_position_error));
     location.velocity =
         VelocityShape::horizontal(vtg->speed_over_ground(), vtg->true_course_over_ground());
 
@@ -169,7 +188,7 @@ PLI_Result provide_location_information_callback_fake(LocationInformation&  loca
     location.time     = TAI_Time::now();
     location.location = LocationShape::ha_ellipsoid_altitude_with_uncertainty(
         options->latitude, options->longitude, options->altitude,
-        HorizontalAccuracy::from_ellipse(0.5, 0.5, 0), VerticalAccuracy::from_1sigma(0.5));
+        HorizontalAccuracy::to_ellipse_39(0.5, 0.5, 0), VerticalAccuracy::from_1sigma(0.5));
 
     location.velocity = VelocityShape::horizontal_vertical_with_uncertainty(10, 0.5, 90, 1, 0.5,
                                                                             VerticalDirection::Up);