+++ examples/lpp/osr_example.h
@@ -10,6 +10,7 @@ enum class Format {
 #endif
     XER,
     ASN1_UPER,
+    NONE,
 };
 
 enum class MsmType {
@@ -24,12 +25,18 @@ class OsrCommand final : public Command {
 public:
     OsrCommand()
         : Command("osr", "Request observation data from a location server"), mFormatArg(nullptr),
-          mMsmTypeArg(nullptr), mPrintRTCMArg(nullptr) {}
+          mLRFMessageIdArg(nullptr), mMsmTypeArg(nullptr), mPrintRTCMArg(nullptr), mNoGPS(nullptr),
+          mNoGLONASS(nullptr), mNoGalileo(nullptr), mNoBeiDou(nullptr) {}
 
     ~OsrCommand() override {
         delete mFormatArg;
+        delete mLRFMessageIdArg;
         delete mMsmTypeArg;
         delete mPrintRTCMArg;
+        delete mNoGPS;
+        delete mNoGLONASS;
+        delete mNoGalileo;
+        delete mNoBeiDou;
     }
 
     void parse(args::Subparser& parser) override;
@@ -37,8 +44,13 @@ public:
 
 private:
     args::ValueFlag<std::string>* mFormatArg;
+    args::ValueFlag<int>*         mLRFMessageIdArg;
     args::ValueFlag<std::string>* mMsmTypeArg;
     args::Flag*                   mPrintRTCMArg;
+    args::Flag*                   mNoGPS;
+    args::Flag*                   mNoGLONASS;
+    args::Flag*                   mNoGalileo;
+    args::Flag*                   mNoBeiDou;
 };
 
 }  // namespace osr_example